jsRender

+++ assets/js/task.js
@@ -20,24 +20,52 @@
   }
 
   function removeFileConfirm (reply, id, name) {
-    $('#confirmModalTitle').text(jsLocale.task.removefile.title)
-    $('#confirmModalBody').text(jsLocale.task.removefile.body(name))
-    $('#confirmModalButtons').html('<button type="button" class="btn btn-danger fileRemoveButton" data-removereply="' + reply + '" data-removeid="' + id + '">' + jsLocale.task.removefile.button + '</button>')
-    $('#confirmModal').modal()
+    let removeModal = $('.removeModal')
+
+    if (removeModal.length > 0) {
+      return false
+    }
+
+    const tmplRemoveModal = $.templates('#tmplRemoveModal')
+
+    const renderedRemoveModal = tmplRemoveModal.render({name: name, reply: reply, id: id})
+
+    $('#modalContainer').html(renderedRemoveModal)
+
+    removeModal = $('#removeModal')
+    removeModal.modal()
+
+    removeModal.on('hidden.bs.modal', function () {
+      $(this).remove()
+    })
+
     $('.fileRemoveButton').on('click', function () {
       const reply = $(this).data('removereply')
       const id = $(this).data('removeid')
       removeFile(reply, id)
+      removeModal.remove()
     })
   }
 
   function replaceFile (reply, id, name, ext) {
-    $('#confirmModalTitle').text(jsLocale.task.replacefile.title(name))
-    $('#confirmModalBody').html('<form action="reply/' + reply + '/updateFile/' + id + '" enctype="multipart/form-data" method="post">' +
-      '<input type="file" name="file" accept=".' + ext + '">' +
-      '<input type="submit" value="' + jsLocale.task.replacefile.button + '" plik" class="btn btn-primary"/>')
-    $('#confirmModalButtons').html('')
-    $('#confirmModal').modal()
+    let replaceModal = $('.replaceModal')
+
+    if (replaceModal.length > 0) {
+      return false
+    }
+    const tmplReplaceModal = $.templates('#tmplReplaceModal')
+
+    const renderedReplaceModal = tmplReplaceModal.render({name: name, reply: reply, id: id, ext: ext})
+
+    $('#modalContainer').html(renderedReplaceModal)
+
+    replaceModal = $('#replaceModal')
+
+    replaceModal.modal()
+
+    replaceModal.on('hidden.bs.modal', function () {
+      $(this).remove()
+    })
   }
   $('.replyFileIcon').on('click', function () {
     const reply = $(this).data('filereply')
@@ -56,7 +84,7 @@
         } else if (data.mimeType === 'image/bmp') {
           $('#fileContentModalBody').html('<img class="img-fluid" src="data:image/bmp;base64,' + data.body + '"/>')
         } else {
-          $('#fileContentModalBody').text(jsLocale.task.showfile.unsupported)
+          $('#fileContentModalBody').text('bad file ext') // TODO: i18n this
         }
         if (!replySent) {
           $('#fileContentModalRemove').on('click', function () {
@@ -73,7 +101,7 @@
         $('#fileContentModal').modal()
       })
       .fail(function (jqXHR, textStatus, errorThrown) {
-        alert('Nie udało się wczytać danych: ' + textStatus + ' - ' + errorThrown + (jqXHR.responseJSON ? '\n' + jqXHR.responseJSON : ''))
+        alert('Nie udało się wczytać danych: ' + textStatus + ' - ' + errorThrown + (jqXHR.responseJSON ? '\n' + jqXHR.responseJSON : '')) // TODO: this into modal
       })
   }
 
@@ -101,20 +129,32 @@
     })
   }
 
-  function renderComment (user, comment, date, read) {
-    let template = $('#commentAjaxTemplate').find('.list-group-item').clone()
-    template.attr('id', 'commentFadeIn')
-    if (user) {
-      let append = user.isTeacher ? '<span class="badge badge-primary">Prowadzący</span><span>&nbsp;napisał(a):</span>' : '<span>&nbsp;napisał(a):</span>'
-      template.find('.task-c-author').text(user.name + ' ' + user.surname).append(append)
-    } else {
-      template.find('.task-c-author').html('<span class="badge badge-info">Wiadomość systemowa</span>')
-    }
-    template.find('.task-c-timestamp').text(date)
-    template.find('.task-c-comment').text(comment)
-    template.find('.task-c-read').html((read ? commenticons.read : commenticons.unread))
+  function renderComment (user, commentContent, date, read) {
+    const tmplCommentAjax = $.templates('#tmplCommentAjax')
 
-    $(template).insertBefore(markAsReadButton)
+    const template = tmplCommentAjax.render({
+      user: (!user ? false : user),
+      teacher: (user ? user.isTeacher : false),
+      name: (user ? user.name + ' ' + user.surname : ''),
+      comment: commentContent,
+      timestamp: date,
+      read: read
+    })
+
+    markAsReadButton.before(template)
+
+    const comment = $('#commentFadeIn')
+
+    // // let template = $('#commentAjaxTemplate').find('.list-group-item').clone()
+    // if (user) {
+    //   let append = user.isTeacher ? '<span class="badge badge-primary">Prowadzący</span><span>&nbsp;napisał(a):</span>' : '<span>&nbsp;napisał(a):</span>'
+    //   comment.find('.task-c-author').text(user.name + ' ' + user.surname).append(append)
+    // } else {
+    //   comment.find('.task-c-author').html('<span class="badge badge-info">Wiadomość systemowa</span>')
+    // }
+    // comment.find('.task-c-timestamp').text(date)
+    // comment.find('.task-c-comment').text(commentContent)
+    comment.find('.task-c-read').html((read ? commenticons.read : commenticons.unread))
 
     if (read === false) {
       markAsReadButton.fadeIn()
@@ -125,9 +165,8 @@
     //   $('#commentArea').append(template)
     // }
 
-    const commentFadeIn = $('#commentFadeIn')
-    commentFadeIn.fadeIn()
-    commentFadeIn.attr('id', '')
+    comment.fadeIn()
+    comment.attr('id', '')
   }
 
   function sendComment (topic, task) {
@@ -189,10 +228,26 @@
     setTimeout(checkComments, 600)
   })
   $('#sendReply').on('click', function () {
-    $('#confirmModalTitle').text('Wysłanie rozwiązania')
-    $('#confirmModalBody').html('Czy na pewno chcesz wysłać rozwiązanie? <br><b>Późniejsza edycja będzie niemożliwa!</b>')
-    $('#confirmModalButtons').html(`<a href='/topic/` + $(this).data('topic') + `/task/` + $(this).data('task') + `/sendReply/' class='btn btn-primary'>Wyślij rozwiązanie</a>`)
-    $('#confirmModal').modal()
+    let sendReplyModal = $('.sendReplyModal')
+
+    if (sendReplyModal.length > 0) {
+      return false
+    }
+
+    const tmplSendReplyModal = $.templates('#tmplSendReplyModal')
+
+    const renderedSendReplyModal = tmplSendReplyModal.render({topic: $(this).data('topic'), task: $(this).data('task')})
+
+    $('#modalContainer').html(renderedSendReplyModal)
+
+    sendReplyModal = $('#sendReplyModal')
+    sendReplyModal.modal()
+
+    sendReplyModal.on('hidden.bs.modal', function () {
+      $(this).remove()
+    })
+
+    sendReplyModal.modal()
   })
 
   $(markAsReadButton).on('click', function () {
